requirements.txt
# Web Framework
Flask==3.1.0
Werkzeug==3.1.0

# HTTP Requests
requests==2.32.3

# Image Processing
Pillow==11.0.0

# Production WSGI Server
gunicorn==23.0.0

# Environment Configuration
python-dotenv==1.0.1

# Input Validation
Flask-Limiter==3.8.0

# Testing (development)
pytest==8.3.4
pytest-cov==6.0.0
